+++ include/chemfiles/capi/atom.h
@@ -180,7 +180,7 @@ CHFL_EXPORT chfl_status chfl_atom_covalent_radius(
 /// Get the atomic number of an `atom` from the atom type, in the integer
 /// pointed to by `number`.
 ///
-/// If the atomic number in unknown, this function set `number` to UINT64_MAX.
+/// If the atomic number in unknown, this function set `number` to 0.
 ///
 /// @example{tests/capi/doc/chfl_atom/atomic_number.c}
 /// @return The operation status code. You can use `chfl_last_error` to learn