+++ src/capi/cell.cpp
@@ -135,11 +135,6 @@ extern "C" chfl_status chfl_cell_set_shape(CHFL_CELL* const cell, chfl_cellshape
     )
 }
 
-extern "C" chfl_status chfl_cell_free(CHFL_CELL* const cell) {
-    delete cell;
-    return CHFL_SUCCESS;
-}
-
 extern "C" chfl_status chfl_cell_wrap(const CHFL_CELL* const cell, chfl_vector3d vector) {
     CHECK_POINTER(cell);
     CHECK_POINTER(vector);
@@ -149,4 +144,9 @@ extern "C" chfl_status chfl_cell_wrap(const CHFL_CELL* const cell, chfl_vector3d
         vector[1] = result[1];
         vector[2] = result[2];
     )
+}
+
+extern "C" chfl_status chfl_cell_free(CHFL_CELL* const cell) {
+    delete cell;
+    return CHFL_SUCCESS;
 }