+++ tests/capi/doc/chfl_residue/with_id.c
@@ -3,17 +3,16 @@
 
 #include <chemfiles.h>
 #include <stdlib.h>
-#include <assert.h>
 
 int main() {
     // [example]
-    CHFL_TOPOLOGY* topology = chfl_topology();
-    CHFL_RESIDUE* residue = chfl_residue("res");
+    CHFL_RESIDUE* residue = chfl_residue_with_id("water", 3);
 
-    chfl_topology_add_residue(topology, residue);
+    if (residue == NULL) {
+        /* handle error */
+    }
 
     chfl_residue_free(residue);
-    chfl_topology_free(topology);
     // [example]
     return 0;
 }